Move UUID implementation from tests to arrow-vector package
