+++ vector/src/main/java/org/apache/arrow/vector/complex/impl/UuidWriterFactory.java
@@ -14,10 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.arrow.vector.holder;
+package org.apache.arrow.vector.complex.impl;
 
-import org.apache.arrow.vector.holders.ExtensionHolder;
+import org.apache.arrow.vector.ExtensionTypeVector;
+import org.apache.arrow.vector.extension.UuidVector;
 
-public class UuidHolder extends ExtensionHolder {
-  public byte[] value;
+public class UuidWriterFactory implements ExtensionTypeWriterFactory {
+
+  @Override
+  public AbstractFieldWriter getWriterImpl(ExtensionTypeVector extensionTypeVector) {
+    if (extensionTypeVector instanceof UuidVector) {
+      return new UuidWriterImpl((UuidVector) extensionTypeVector);
+    }
+    return null;
+  }
 }
