+++ vector/src/main/java/org/apache/arrow/vector/complex/impl/UuidWriterImpl.java
@@ -18,8 +18,8 @@ package org.apache.arrow.vector.complex.impl;
 
 import java.nio.ByteBuffer;
 import java.util.UUID;
-import org.apache.arrow.vector.UuidVector;
-import org.apache.arrow.vector.holder.UuidHolder;
+import org.apache.arrow.vector.extension.UuidHolder;
+import org.apache.arrow.vector.extension.UuidVector;
 import org.apache.arrow.vector.holders.ExtensionHolder;
 
 public class UuidWriterImpl extends AbstractExtensionTypeWriter<UuidVector> {
